+++ netanalysis/dns/google/get_google_dns_locations.sh
